Move code to testFixtures
- #19867
