+++ subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r68/AbstractKotlinDslScriptsModelCrossVersionSpec.groovy
@@ -17,6 +17,7 @@
 package org.gradle.kotlin.dsl.tooling.builders.r68
 
 import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
+import org.gradle.kotlin.dsl.tooling.builders.BuildSpec
 import org.gradle.tooling.model.kotlin.dsl.KotlinDslScriptsModel
 
 import static org.gradle.integtests.tooling.fixture.TextUtil.escapeString
