+++ subprojects/kotlin-dsl-tooling-builders/src/testFixtures/groovy/org/gradle/kotlin/dsl/tooling/builders/ProjectSourceRoots.groovy
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.kotlin.dsl.tooling.builders
+
+
+
+
+class ProjectSourceRoots {
+
+    final File projectDir
+    final List<String> sourceSets
+    final List<String> languages
+
+    ProjectSourceRoots(File projectDir, List<String> sourceSets, List<String> languages) {
+        this.projectDir = projectDir
+        this.sourceSets = sourceSets
+        this.languages = languages
+    }
+}
